+++ src/components/AsideHeader/components/FirstPanel.tsx
@@ -25,6 +25,7 @@ export const FirstPanel = React.forwardRef<HTMLDivElement>((_props, ref) => {
         customBackgroundClassName,
         className,
         hideCollapseButton,
+        logo,
         qa,
     } = useAsideHeaderInnerContext();
     const visibleMenuItems = useVisibleMenuItems();
@@ -46,7 +47,8 @@ export const FirstPanel = React.forwardRef<HTMLDivElement>((_props, ref) => {
                             {customBackground}
                         </div>
                     )}
-                    <Header />
+                    {/* TODO add new prop for hiding Header */}
+                    {logo && <Header />}
                     {visibleMenuItems?.length ? (
                         <CompositeBar
                             type="menu"