+++ src/components/Footer/mobile/__tests__/helpersPlaywright.ts
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as DefaultMobileFooterStories from '../__stories__/MobileFooter.stories';
+
+export const MobileFooterStories = composeStories(DefaultMobileFooterStories);