+++ src/components/HotkeysPanel/__tests__/helpersPlaywright.ts
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as DefaultHotkeysPanelStories from '../__stories__/HotkeysPanel.stories';
+
+export const HotkeysPanelStories = composeStories(DefaultHotkeysPanelStories);