+++ src/components/Logo/__tests__/Logo.visual.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import {test} from '~playwright/core';
+
+import {LogoStories} from './helpersPlaywright';
+
+test.describe('Logo', () => {
+    test('render story: <Showcase>', async ({mount, expectScreenshot}) => {
+        await mount(<LogoStories.Showcase />);
+
+        await expectScreenshot();
+    });
+});