+++ src/components/Logo/__tests__/helpersPlaywright.ts
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as DefaultLogoStories from '../__stories__/Logo.stories';
+
+export const LogoStories = composeStories(DefaultLogoStories);