+++ src/components/MobileHeader/BurgerMenu/__tests__/helpersPlaywright.ts
@@ -1,0 +1,5 @@
+import {composeStories} from '@storybook/react';
+
+import * as DefaultBurgerMenuStories from '../__stories__/BurgerMenu.stories';
+
+export const BurgerMenuStories = composeStories(DefaultBurgerMenuStories);